requirements.txt
python-dotenv>=1.0.0
openai==1.3.7
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
moviepy>=1.0.3
Google-Images-Search==1.4.6
elevenlabs==0.2.27
python-youtube==0.9.0
pytiktokapi==0.0.5
duckduckgo-search==4.1.1